Add Tacticboard Routes Controller
